requirements.txt
-i https://pypi.org/simple
algoliasearch-django==1.7.1
algoliasearch==1.20.0
boto3==1.9.190
botocore==1.12.190
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
django-allauth==0.39.1
django-appconf==1.0.3
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-extensions==2.2.1
django-jet==1.0.8
django-js-asset==1.2.2
django-location-field==2.1.0
django-rest-framework==0.1.0
django-select2==7.1.0
django-stdimage==4.1.0
django-storages==1.7.1
django==2.2.3
djangorestframework==3.10.1
docutils==0.14
idna==2.8
jmespath==0.9.4
oauthlib==3.0.2
pillow==6.1.0
progressbar2==3.42.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0 ; python_version >= '2.7'
python-utils==2.3.0
python3-openid==3.1.0
pytz==2019.1
requests-oauthlib==1.2.0
requests==2.22.0
rollbar==0.14.7
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3 ; python_version >= '3.4'
